migrate denonmarantz and JDK11 compatibility
