+++ CODEOWNERS
@@ -25,7 +25,6 @@
 /addons/binding/org.openhab.binding.coolmasternet/ @projectgus
 /addons/binding/org.openhab.binding.daikin/ @caffineehacker
 /addons/binding/org.openhab.binding.deconz/ @davidgraeff
-/addons/binding/org.openhab.binding.denonmarantz/ @jwveldhuis
 /addons/binding/org.openhab.binding.digitalstrom/ @MichaelOchel @msiegele
 /addons/binding/org.openhab.binding.dlinksmarthome/ @MikeJMajor
 /addons/binding/org.openhab.binding.dscalarm/ @RSStephens
@@ -228,6 +227,7 @@
 /bundles/org.openhab.binding.astro/ @gerrieg
 /bundles/org.openhab.binding.avmfritz/ @cweitkamp
 /bundles/org.openhab.binding.bigassfan/ @mhilbush
+/bundles/org.openhab.binding.denonmarantz/ @jwveldhuis
 /bundles/org.openhab.binding.dmx/ @J-N-K
 /bundles/org.openhab.binding.onewire/ @J-N-K
 /bundles/org.openhab.binding.openweathermap/ @cweitkamp
