--- addons/binding/org.openhab.binding.denonmarantz/META-INF/MANIFEST.MF
@@ -1,35 +1,0 @@
-Manifest-Version: 1.0
-Automatic-Module-Name: org.openhab.binding.denonmarantz
-Bundle-ActivationPolicy: lazy
-Bundle-ClassPath: .
-Bundle-ManifestVersion: 2
-Bundle-Name: DenonMarantz Binding
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-SymbolicName: org.openhab.binding.denonmarantz;singleton:=true
-Bundle-Vendor: openHAB
-Bundle-Version: 2.5.0.qualifier
-Import-Package: 
- javax.jmdns,
- javax.xml.bind,
- javax.xml.bind.annotation,
- javax.xml.bind.annotation.adapters,
- org.apache.commons.io,
- org.apache.commons.lang,
- org.apache.commons.net,
- org.eclipse.jdt.annotation;resolution:=optional,
- org.eclipse.jetty.client,
- org.eclipse.jetty.client.api,
- org.eclipse.jetty.http,
- org.eclipse.jetty.util.component,
- org.eclipse.smarthome.config.core,
- org.eclipse.smarthome.config.discovery,
- org.eclipse.smarthome.config.discovery.mdns,
- org.eclipse.smarthome.core.library.types,
- org.eclipse.smarthome.core.thing,
- org.eclipse.smarthome.core.thing.binding,
- org.eclipse.smarthome.core.thing.binding.builder,
- org.eclipse.smarthome.core.thing.type,
- org.eclipse.smarthome.core.types,
- org.eclipse.smarthome.io.net.http,
- org.slf4j
-Service-Component: OSGI-INF/*.xml
