--- addons/binding/org.openhab.binding.denonmarantz/OSGI-INF/.gitignore
@@ -1,2 +1,0 @@
-*.xml
-
