--- addons/binding/org.openhab.binding.denonmarantz/build.properties
@@ -1,8 +1,0 @@
-source.. = src/main/java/
-output.. = target/classes
-bin.includes = META-INF/,\
-               .,\
-               OSGI-INF/,\
-               ESH-INF/,\
-               NOTICE
-
