+++ bom/openhab-addons/pom.xml
@@ -127,7 +127,7 @@
       <version>${project.version}</version>
     </dependency>
     <dependency>
-      <groupId>org.openhab.binding</groupId>
+      <groupId>org.openhab.addons.bundles</groupId>
       <artifactId>org.openhab.binding.denonmarantz</artifactId>
       <version>${project.version}</version>
     </dependency>
