+++ bundles/org.openhab.binding.denonmarantz/src/main/resources/ESH-INF/binding/binding.xml
