+++ bundles/org.openhab.binding.denonmarantz/src/main/resources/ESH-INF/thing/thing-types.xml
