+++ bundles/pom.xml
@@ -23,6 +23,7 @@
     <module>org.openhab.binding.astro</module>
     <module>org.openhab.binding.avmfritz</module>
     <module>org.openhab.binding.bigassfan</module>
+    <module>org.openhab.binding.denonmarantz</module>
     <module>org.openhab.binding.dmx</module>
     <module>org.openhab.binding.onewire</module>
     <module>org.openhab.binding.openweathermap</module>
